backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
fpdf2
python-docx
Pillow
pydantic
PyMuPDF
